[feat] refactors views; makes footer data consistent across views
