+++ site/components/LandscapeOverlay/LandscapeOverlay.tsx
@@ -7,6 +7,7 @@ import { River, rivers } from "@/simulation/data/rivers";
 import Link from "next/link";
 import PointerLockButton from "../PointerLockButton/PointerLockButton";
 import RiverLocations from "../RiverLocations/RiverLocations";
+import FooterRow from "../FooterRow/FooterRow";
 
 interface DroneViewOverlayProps {
   river: River;
@@ -14,20 +15,6 @@ interface DroneViewOverlayProps {
   t: number;
   w: number;
 }
-
-const formatAsYears = (t: number): string => {
-  return `${t.toFixed(0)} steps`;
-};
-
-const formatAsLatLong = (t: [number, number]): string => {
-  return `${t[0].toFixed(5)}, ${t[1].toFixed(5)}`;
-};
-
-const formatAsVolume = (w: number): string => {
-  return `${Math.round(w).toLocaleString("es-MX")} / ${(
-    512 * 512
-  ).toLocaleString("es-MX")} cells`;
-};
 
 export default function DroneViewOverlay({
   river,
@@ -68,34 +55,18 @@ export default function DroneViewOverlay({
           onClick={() => setOverlayState(UIOverlayState.SimTools)}
         >
           <OverlayButton>
-            <span>Tools</span>
+            <span>Simulation</span>
           </OverlayButton>
         </div>
       </div>
 
-      <div className="flex mt-auto w-full items-left">
-        <div className="text-[#ff0000] flex mr-auto">
-          <div>
-            <OverlayButton>
-              <span>Info</span>
-            </OverlayButton>
-          </div>
-        </div>
-        <div className="ml-auto flex text-right">
-          <div className="px-10 py-1">
-            {formatAsLatLong(river.coordinates)}{" "}
-          </div>
-          <div className="px-10 py-1">{formatAsVolume(w)} </div>
-          <div className="px-10 py-1">{formatAsYears(t)}</div>
-
-          {/* <div className="text-white ml-2">Test</div> */}
-        </div>
-        <div onClick={() => setRunning(!isRunning)} className="w-32 text-right">
-          <OverlayButton>
-            {isRunning ? <span>Pause</span> : <span>Run</span>}
-          </OverlayButton>
-        </div>
-      </div>
+      <FooterRow
+        river={river}
+        t={t}
+        w={w}
+        isRunning={isRunning}
+        setRunning={setRunning}
+      />
     </div>
   );
 }
